# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "platform API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {}
  },
  "paths": {
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "institution": {
                        "type": "string"
                      },
                      "kind": {
                        "type": "string",
                        "enum": [
                          "depository",
                          "credit",
                          "loan",
                          "investment",
                          "other"
                        ]
                      },
                      "subtype": {
                        "type": "string"
                      },
                      "mask": {
                        "type": "string"
                      },
                      "balance": {
                        "type": "object",
                        "properties": {
                          "amountMinor": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "amountMinor",
                          "currency"
                        ],
                        "additionalProperties": false
                      },
                      "creditLimit": {
                        "type": "object",
                        "properties": {
                          "amountMinor": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "amountMinor",
                          "currency"
                        ],
                        "additionalProperties": false
                      },
                      "balanceAsOf": {
                        "type": "string"
                      },
                      "isActive": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "institution",
                      "kind",
                      "balance",
                      "isActive"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "to",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "accountId": {
                        "type": "string"
                      },
                      "postedAt": {
                        "type": "string"
                      },
                      "authorizedAt": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "merchant": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "object",
                        "properties": {
                          "amountMinor": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "amountMinor",
                          "currency"
                        ],
                        "additionalProperties": false
                      },
                      "pending": {
                        "type": "boolean"
                      },
                      "category": {
                        "type": "string"
                      },
                      "source": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "accountId",
                      "postedAt",
                      "description",
                      "amount",
                      "pending",
                      "category",
                      "source"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "kind": {
                        "type": "string",
                        "enum": [
                          "savings_net_flow",
                          "balance_target",
                          "debt_paydown"
                        ]
                      },
                      "accountId": {
                        "type": "string"
                      },
                      "targetAmount": {
                        "type": "object",
                        "properties": {
                          "amountMinor": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "amountMinor",
                          "currency"
                        ],
                        "additionalProperties": false
                      },
                      "targetDate": {
                        "type": "string"
                      },
                      "startedAt": {
                        "type": "string"
                      },
                      "baselineAmount": {
                        "type": "object",
                        "properties": {
                          "amountMinor": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "amountMinor",
                          "currency"
                        ],
                        "additionalProperties": false
                      },
                      "note": {
                        "type": "string"
                      },
                      "active": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "kind",
                      "targetAmount",
                      "active"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/budgets": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "category": {
                        "type": "string"
                      },
                      "monthlyCap": {
                        "type": "object",
                        "properties": {
                          "amountMinor": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "amountMinor",
                          "currency"
                        ],
                        "additionalProperties": false
                      },
                      "active": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "category",
                      "monthlyCap",
                      "active"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asOf": {
                      "type": "string"
                    },
                    "metrics": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "asOf",
                    "metrics"
                  ],
                  "additionalProperties": false,
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investments/activity": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "to",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "accountId": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "kind": {
                        "type": "string",
                        "enum": [
                          "contribution",
                          "dividend",
                          "interest",
                          "buy",
                          "sell",
                          "other"
                        ]
                      },
                      "amount": {
                        "type": "object",
                        "properties": {
                          "amountMinor": {
                            "type": "integer"
                          },
                          "currency": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "amountMinor",
                          "currency"
                        ],
                        "additionalProperties": false
                      },
                      "ticker": {
                        "type": "string"
                      },
                      "quantity": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "accountId",
                      "date",
                      "description",
                      "kind",
                      "amount"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights/history": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "from",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "in": "query",
            "name": "to",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "asOf": {
                        "type": "string"
                      },
                      "metrics": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "asOf",
                      "metrics"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "since",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "occurredOn": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "payload": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "type",
                      "occurredOn",
                      "createdAt",
                      "payload"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}